requirements.txt
torch>=2.0.0
torchvision>=0.15.0
efficientnet-pytorch>=0.7.1
streamlit>=1.28.0
shap>=0.43.0
opencv-python>=4.8.0
reportlab>=4.0
Pillow>=10.0.0
matplotlib>=3.7.0
numpy>=1.24.0
pandas>=2.0.0
